[IMP] queue_management_display
