+++ .copier-answers.yml
@@ -1,8 +1,8 @@
 # Do NOT update manually; changes here will be overwritten by Copier
-_commit: v1.14.1
+_commit: v1.20
 _src_path: https://github.com/OCA/oca-addons-repo-template.git
 ci: GitHub
-dependency_installation_mode: PIP
+convert_readme_fragments_to_markdown: false
 generate_requirements_txt: true
 github_check_license: true
 github_ci_extra_env: {}
@@ -11,6 +11,7 @@ github_enable_makepot: false
 github_enable_stale_action: true
 github_enforce_dev_status_compatibility: true
 include_wkhtmltopdf: false
+odoo_test_flavor: Both
 odoo_version: 14.0
 org_name: CreuBlanca
 org_slug: tegin
@@ -19,6 +20,6 @@ repo_description: Module to manage queues
 repo_name: Queue Management Module
 repo_slug: kiwi
 repo_website: https://github.com/tegin/kiwi
-travis_apt_packages: []
-travis_apt_sources: []
+use_pyproject_toml: false
+use_ruff: false
