+++ queue_management_display/static/src/js/notification/queue_management_view.js
@@ -1,23 +1,22 @@
-odoo.define("queue_management.QueueDisplayControlView", function (require) {
+odoo.define("queue_management.QueueDisplayNotificationView", function (require) {
     "use strict";
 
     var BasicView = require("web.BasicView");
-    var QueueDisplayControlController = require("queue_management.QueueDisplayControlController");
-    // Var DashboardModel = require("kpi_dashboard.DashboardModel");
-    var QueueDisplayControlRenderer = require("queue_management.QueueDisplayControlRenderer");
+    var QueueDisplayNotificationController = require("queue_management.QueueDisplayNotificationController");
+    var QueueDisplayNotificationRenderer = require("queue_management.QueueDisplayNotificationRenderer");
     var view_registry = require("web.view_registry");
     var core = require("web.core");
 
     var _lt = core._lt;
 
-    var QueueDisplayControlView = BasicView.extend({
+    var QueueDisplayNotificationView = BasicView.extend({
         accesskey: "p",
         display_name: _lt("Display"),
         icon: "fa-tachometer",
-        viewType: "queue_display_control",
+        viewType: "queue_display_notification",
         config: _.extend({}, BasicView.prototype.config, {
-            Controller: QueueDisplayControlController,
-            Renderer: QueueDisplayControlRenderer,
+            Controller: QueueDisplayNotificationController,
+            Renderer: QueueDisplayNotificationRenderer,
             // Model: DashboardModel,
         }),
         multi_record: false,
@@ -33,7 +32,7 @@ odoo.define("queue_management.QueueDisplayControlView", function (require) {
         },
     });
 
-    view_registry.add("queue_display_control", QueueDisplayControlView);
+    view_registry.add("queue_display_notification", QueueDisplayNotificationView);
 
-    return QueueDisplayControlView;
+    return QueueDisplayNotificationView;
 });
