+++ queue_management_display/views/queue_display.xml
@@ -35,12 +35,16 @@
                         <field name="show_items" />
                         <field name="max_time" widget="float_time" />
                         <field name="shiny_time" widget="float_time" />
-                        <field
-                            name="location_ids"
-                            context="{'tree_view_ref':'queue_management.queue_location_config_tree_view', 'form_view_ref': 'queue_management.queue_location_config_form_view'}"
-                        />
+                        <field name="audio_file" filename="audio_filename" />
+                        <field name="audio_filename" invisible="1" />
                     </group>
                     <notebook>
+                        <page name="locations" string="Locations">
+                            <field
+                                name="location_ids"
+                                context="{'tree_view_ref':'queue_management.queue_location_config_tree_view', 'form_view_ref': 'queue_management.queue_location_config_form_view'}"
+                            />
+                        </page>
                         <page name="qweb" string="QWeb">
                             <field name="qweb" widget="ace" options="{'mode': 'xml'}" />
                         </page>
@@ -67,6 +71,7 @@
                 <field name="shiny_time" />
                 <field name="qweb" />
                 <field name="css" />
+                <field name="audio_file" />
             </form>
         </field>
     </record>
@@ -106,7 +111,7 @@
         <field name="view_mode">form</field>
         <field name="target">fullscreen</field>
         <field name="domain">[]</field>
-        <field name="context">{}</field>
+        <field name="context">{'bin_size': 0}</field>
     </record>
 
     <record
